If a package is already downloaded in the cache, skip calling 'go get' in the cache
Summary:
This fixes a tricky bug that @timotyen found. If a package in the cache is not compilable, or if it
imports a package that isn't go-gettable, then any call to `retool add` or `retool upgrade that
package would fail because the package would fail on a (unnecessary!) `go get` call.

Users could get in this state by adding a package at a particular, bad commit. Retool uses `git
checkout` to switch to that commit, not `go get`, so it was possible to move into a bad commit, but
not possible to get out of one.

Also, this is the first commit to add a test. Yay!

Test Plan: `go test` \o/

Reviewers: timotyen

Subscribers: timotyen, kelleyk

Differential Revision: https://phabricator.internal.justin.tv/D4958