Fix gometalinter (#15)
* Set GOPATH and GOBIN in retool do

This is required for gometalinter support, and makes sense to do anyway in
general. This convinces tools to use the vendored _tools directory.

* Don't delete gometalinter's vendored linters

This is a special case, but gometalinter is so widely used that it's worth
doing.

* Don't overwrite GOPATH, just append to it

Overwriting GOPATH confuses any tools that use the go/build package to analyze
source code. That includes most linters. Instead of totally overwriting it, we
should just append _tools as another possible GOPATH.

* Trim down complexity of do.go